// RMS-BackOfficePortal/src/app/core/models/ri-pl.interface.ts
export interface RIPL {

    i_txn_type: string;
    i_entity_type: string;
    i_entity_no: string;
    i_calendar_yr: Date;
    i_dt_due: Date;
    i_dt_impair: Date;
    i_dt_writeoff: Date;
    i_ripl_ctype: string;
    i_accr_amt: number;
    i_rcpt_no: string;
    i_dt_created: Date;
    i_status: string;
    

}